requirement.txt
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
Babel==2.12.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crispy-bootstrap5==0.7
cryptography==41.0.4
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.1.13
django-allauth==0.56.1
django-appconf==1.0.5
django-binary-database-files==1.0.18
django-compressor==4.4
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-storages==1.14.2
dnspython==2.6.0
docxcompose==1.4.0
filelock==3.12.4
flatbuffers==23.5.26
gunicorn==20.1.0
html5lib==1.1
humanfriendly==10.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
Jinja2==3.1.2
lazy_loader==0.3
llvmlite==0.41.0
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
oscrypto==1.3.0
packaging==23.1
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.3
psycopg2-binary==2.9.5
pycparser==2.21
pyee==8.2.2
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
PyMatting==1.1.8
pymongo==4.6.1
pypdf==3.16.1
pypng==0.20220715.0
pyreadline3==3.4.1
python-bidi==0.4.2
python-docx==0.8.11
python3-openid==3.2.0
PyWavelets==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.1
redis==5.0.1
rembg==2.0.50
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
scikit-image==0.21.0
scipy==1.11.2
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.2.4
svglib==1.5.1
sympy==1.12
tifffile==2023.9.18
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==1.26.16
webencodings==0.5.1
websocket-client==1.6.3
websockets==10.4
whitenoise==6.4.0
xhtml2pdf==0.2.11
zipp==3.17.0
